refactor: move check command inside subfolder
