+++ cmd/check/init.go
@@ -1,7 +1,7 @@
 /*
 Copyright © 2025 NAME HERE <EMAIL ADDRESS>
 */
-package cmd
+package check
 
 import (
 	"github.com/opt-nc/geol/v2/cmd/templates"
@@ -13,8 +13,8 @@ var force bool
 var appName string
 var appID string
 
-// initCmd represents the init command
-var initCmd = &cobra.Command{
+// InitCmd represents the init command
+var InitCmd = &cobra.Command{
 	Use:     "init",
 	Aliases: []string{"i"},
 	Short:   "Generate a valid template configuration file for check command",
@@ -31,9 +31,8 @@ geol check init --output stack.yaml --force`,
 }
 
 func init() {
-	checkCmd.AddCommand(initCmd)
-	initCmd.Flags().StringVarP(&output, "output", "o", ".geol.yaml", "Path to the output file")
-	initCmd.Flags().BoolVarP(&force, "force", "f", false, "Overwrite the file if it already exists")
-	initCmd.Flags().StringVarP(&appName, "app-name", "a", "", "Application name to use in the generated template")
-	initCmd.Flags().StringVar(&appID, "app-id", "", "Application ID to use in the generated template")
+	InitCmd.Flags().StringVarP(&output, "output", "o", ".geol.yaml", "Path to the output file")
+	InitCmd.Flags().BoolVarP(&force, "force", "f", false, "Overwrite the file if it already exists")
+	InitCmd.Flags().StringVarP(&appName, "app-name", "a", "", "Application name to use in the generated template")
+	InitCmd.Flags().StringVar(&appID, "app-id", "", "Application ID to use in the generated template")
 }
