+++ cmd/root.go
@@ -13,6 +13,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/opt-nc/geol/v2/cmd/cache"
+	"github.com/opt-nc/geol/v2/cmd/check"
 	"github.com/opt-nc/geol/v2/cmd/exports"
 	"github.com/opt-nc/geol/v2/cmd/list"
 	"github.com/opt-nc/geol/v2/cmd/product"
@@ -91,6 +92,7 @@ func validateWithCue(yamlFile string) error {
 
 func init() {
 	rootCmd.AddCommand(cache.CacheCmd)
+	rootCmd.AddCommand(check.CheckCmd)
 	rootCmd.AddCommand(product.ProductCmd)
 	rootCmd.AddCommand(list.ListCmd)
 	rootCmd.AddCommand(exports.ExportCmd)
